Update android/app/build.gradle to support Release Android APK
- use custom properties file (release-keystore.properties) that contains release keystore credentials
- set signingConfig to use release credentials
- Enable minify codebase
- Setting proguard to improve code minification

release.keystore is used for creating signed Android APK.
It's located in <project_root>/android/app/release.keystore
NOTE: it must NOT be added to version control.
It's to be treated as a secret

release-keystore.properties contains the credentials used for generating
release Android APK.
It's located in <project_root>/android/release-keystore.properties
NOTE: it must NOT be added to version control.
It's to be treated as a secret