Move the stored symbols into a symbol object for Stocks reducer.
The reason for this is because typescript complains when the key can be
any string.
Adding that with 'search' key will cause confusion in typescript.

So moving it to a sub object is important.

NOTE: need to apply changes to other files as well. Likely update the tests as well