Add selectStockQuoteTrim for Stocks selectors
This selector will trim down Stock quote data for each symbol in favorites into basic information.
The data returned will be:
- Stock symbol
- Company name
- Stock price
- price change
- price change percent
- fetching : if the data is being fetched

This is ONLY to be used for the Home screen.

NOTE: a possibility for improvment is to reduce the input array instead
of mapping it. If the data for stock symbol is present, then add it to
the final array. If the data is not present, then skip over.